Make it possible to link to type members directly
Following the path of least resistance: since link anchors are
automatically added to Markdown section headers we can make the
properties each be its own subsection and that gives us anchors
for free, more or less.

We lose the table structure but I'm not sure if I liked the table
in the first place.